make auto `closeDocComment` default; separate jsdoc comments from other document comments
